Extract the Expandable_ListView
